requirements.txt
# ORION (Operational Risk and Investment Optimization Network)
# Python Dependencies

# Web Framework
Flask==2.3.3
Flask-Session==0.5.0

# Database
mysql-connector-python==8.1.0
SQLAlchemy==2.0.20

# Data Science & Analysis
pandas==2.1.0
numpy==1.25.2
yfinance==0.2.28
scipy==1.11.2

# Machine Learning
scikit-learn==1.3.0

# Natural Language Processing
textblob==0.17.1
nltk==3.8.1

# Security
Werkzeug==2.3.7
bcrypt==4.0.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
